review: refactor <MaybePostpone /> component to <Postpone />
